+++ shop/src/main/java/ru/novikov/shop/repository/OrdersToProductsRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import ru.novikov.shop.model.Order;
 import ru.novikov.shop.model.OrdersToProducts;
 
-public interface OrdersToProductsRepository extends JpaRepository<OrdersToProducts, Integer> {
+public interface OrdersToProductsRepository extends JpaRepository<OrdersToProducts, Long> {
     void deleteByOrder(Order order);
 }
