+++ shop/src/main/java/ru/novikov/shop/service/ProductService.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface ProductService {
     Product addProduct(Product product);
-    void delete(int id);
+    void delete(Long id);
     Product getByName(String name);
     Product editProduct(Product product);
     List<Product> getAll();
